requirements.txt
Twisted==13.1.0
Werkzeug==0.9.4
beautifulsoup4==4.3.2
klein==0.2.3
pyasn1==0.1.7
pycrypto==2.6.1
python-keyczar==0.71c
scrypt==0.6.1
requests==2.2.1
zope.interface==4.1.0
